tclInt: Fix TCL_ALIGN for CHERI by not hard-coding 8-byte alignment
This is needed to support CHERI, and thus Arm's experimental Morello
prototype, where pointers are implemented using unforgeable capabilities
that include bounds and permissions metadata to provide fine-grained
spatial and referential memory safety, as well as revocation by sweeping
memory to provide heap temporal memory safety.

These capabilities are twice the size of an integer address, or 16 bytes
on a 64-bit system, and require that same alignment, so 8 bytes is not
sufficient. Adjust the definition of TCL_ALIGN to compute the maximum
alignment needed for long long, double and void *, which should be
sufficient for Tcl's purposes. For backwards compatibility, in case
anything is relying on a minimum of 8-byte alignment (since this is an
exported interface), clamp this to a minimum of 8 so that this change
only affects architectures that require even greater alignment.